Use uniform naming for all controllers
