+++ src/server/controller/StaticController.js
