+++ src/server/routes/posting.js
@@ -4,7 +4,7 @@
  * Routes for all hashtag related requests
  */
 const Router = require('co-router');
-const posting = require('../controller/posting');
+const posting = require('../controller/PostingController');
 
 const router = new Router();
