+++ src/server/routes/static.js
@@ -5,9 +5,9 @@
  */
 const Router = require('co-router');
 const router = new Router();
-const session = require('../controller/session');
+const session = require('../controller/SessionController');
 
-const StaticController = require('../controller/static.js');
+const StaticController = require('../controller/StaticController.js');
 
 router.get('/members', StaticController.renderMemberPage);
